Fix linker error
