+++ barretenberg/cpp/Earthfile
@@ -196,3 +196,4 @@ test:
         ENV HARDWARE_CONCURRENCY=$hardware_concurrency
     END
     RUN cd build && GTEST_COLOR=1 ctest -j$(nproc) --output-on-failure
+
